IMP: Improve how priors are evaluated
Priors are evaluated at a point using the evaluate method. It
would be syntactically much better to simply call the prior to
evaluate it. In addition, a prior should return its gradient
when it is evaluated.

This commit renames the evaluate method of the GaussianPrior class
to __call__ and modifies it to return the value and the gradient.