FEAT: Add gradient norm tolerance as parameter
The cimem.solve function successfully terminates when the norm of the
gradient of the cost function is below a certain threshold. This
threshold is currently hard coded in the function and should be instead
be added as a parameter with a default value.

This commit adds the gradient tolerance as a parameter to the
cimem.solve function.

Closes #3.