FIX: Change default number of states of clusters
Clusters are initialized with 2 states by default. This generates a bug
when the number of priors is not also 2. To fix, the number of states
of a cluster should default to the number of priors supplied.

This commit changes the default number of states to match the number of
priors.